requirements.txt
Django==4.2.23
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
mysqlclient